The Files in the Project
A Puzzle Solver, it can be dynamically coded in with number of columns
and rows edited live, with elements for the colomns and rows edited,
with a seeding image edited.

It also includes File read/write capability using my own take on the CSV
format and text files. A coded in puzzle can be saved for later.

The purpose is to solve the puzzle via algorithms, I can clean it up and
I can definitely improve it but I got it to work in under 20 minutes so
I am content. Python is so very very slow and tkinter is reprehensibly
slow.